+++ app/src/main/java/com/example/peaksrecognition/Config.java
@@ -2,7 +2,7 @@ package com.example.peaksrecognition;
 
 public class Config {
     public double[] initObserverLocation;
-    public double[] initObserverRotation;
+    public float[] initObserverRotation;
     public double maxDistance;
     public double minDistance;
     public float FOVHorizontal;
